refactor (cleanup + merge prepare)
