+++ .gitignore
@@ -1,3 +1,4 @@
+python-hwpx-docs/
 .claude/
 
 # Byte-compiled / optimized / DLL files
