--- LAS-doc-processor/.gitignore
@@ -1,1 +1,0 @@
-python-hwpx-docs/
