+++ apps/backend/doc-processor/.env.template
