+++ apps/backend/doc-processor/.gitignore
@@ -1,0 +1,1 @@
+python-hwpx-docs/
