+++ apps/backend/doc-processor/prompts/categorization.txt
