+++ apps/backend/doc-processor/prompts/prelim_categorization.txt
