+++ apps/backend/doc-processor/tests/doc_samples/(노동)표준근로계약서모음/건설일용직계약서.hwp
