+++ apps/backend/doc-processor/tests/doc_samples/(노동)표준근로계약서모음/연소근로자)표준근로계약서.hwp
