+++ apps/backend/doc-processor/tests/doc_samples/(노동)표준근로계약서모음/친권자동의서.hwp
