+++ apps/backend/doc-processor/tests/doc_samples/(노동)표준근로계약서모음/표준근로계약서(국제 표).hwp
