+++ apps/backend/doc-processor/tests/doc_samples/표준계약서모음(hwp-hwpx)/04. 2차적저작물작성권 양도계약서.hwp
