+++ apps/backend/doc-processor/tests/graph.png
