--- pyproject.toml
@@ -1,19 +1,0 @@
-[project]
-name = "las-system"
-version = "0.1.0"
-description = "Add your description here"
-readme = "README.md"
-requires-python = ">=3.12"
-dependencies = [
-    "httpx>=0.28.1",
-    "jpype1>=1.6.0",
-    "langchain>=1.2.10",
-    "langchain-openai>=1.1.11",
-    "langgraph>=1.0.10",
-    "python-dotenv>=1.2.2",
-    "python-hwpx",
-    "tiktoken>=0.12.0",
-]
-
-[tool.uv.sources]
-python-hwpx = { git = "https://github.com/maxjo0418/python-hwpx.git" }
